Move authenticator to client module.
