+++ src/client/authentication.rs
