+++ src/client/mod.rs
@@ -1,4 +1,6 @@
+mod authentication;
 mod driver;
 
-pub(crate) use driver::REDIRECT_URI;
+pub(crate) const REDIRECT_URI: &str = "https://console.truelayer.com/redirect-page";
+
 pub use driver::{AccountsResult, CardsResult, Environment, TlClient};
